Rename package: featherframework -> feathercore
