+++ protocol/src/main/java/org/feathercore/protocol/Packet.java
@@ -14,13 +14,27 @@
  * limitations under the License.
  */
 
-package org.feathercore.network.packet.status.server;
-
-import org.feathercore.network.packet.status.client.C01Ping;
+package org.feathercore.protocol;
 
 /**
  * Created by k.shandurenko on 09/04/2019
  */
-public class S01Pong extends C01Ping {
+public abstract class Packet implements Cloneable {
+
+    public abstract int getId();
+
+    public abstract void write(Buffer buffer);
+
+    public abstract void read(Buffer buffer);
+
+    @Override
+    public Packet clone() {
+        try {
+            return (Packet) super.clone();
+        } catch (Exception ex) {
+            ex.printStackTrace();
+            return null;
+        }
+    }
 
 }
