+++ protocol/src/main/java/org/feathercore/protocol/packet/status/client/C00ServerQuery.java
@@ -14,27 +14,27 @@
  * limitations under the License.
  */
 
-package org.featherframework.protocol;
+package org.feathercore.protocol.packet.status.client;
+
+import org.feathercore.protocol.Buffer;
+import org.feathercore.protocol.Packet;
 
 /**
  * Created by k.shandurenko on 09/04/2019
  */
-public abstract class Packet implements Cloneable {
-
-    public abstract int getId();
-
-    public abstract void write(Buffer buffer);
-
-    public abstract void read(Buffer buffer);
-
+public class C00ServerQuery extends Packet {
     @Override
-    public Packet clone() {
-        try {
-            return (Packet) super.clone();
-        } catch (Exception ex) {
-            ex.printStackTrace();
-            return null;
-        }
+    public int getId() {
+        return 0x00;
     }
 
+    @Override
+    public void write(Buffer buffer) {
+
+    }
+
+    @Override
+    public void read(Buffer buffer) {
+
+    }
 }
