+++ protocol/src/main/java/org/feathercore/protocol/packet/status/client/C01Ping.java
@@ -14,27 +14,42 @@
  * limitations under the License.
  */
 
-package org.feathercore.network.packet.status.client;
+package org.feathercore.protocol.packet.status.client;
 
-import org.featherframework.protocol.Buffer;
-import org.featherframework.protocol.Packet;
+import org.feathercore.protocol.Buffer;
+import org.feathercore.protocol.Packet;
 
 /**
  * Created by k.shandurenko on 09/04/2019
  */
-public class C00ServerQuery extends Packet {
+public class C01Ping extends Packet {
+
+    private long clientTime;
+
+    public C01Ping(long clientTime) {
+        this.clientTime = clientTime;
+    }
+
+    public C01Ping() {
+    }
+
+    public long getClientTime() {
+        return clientTime;
+    }
+
     @Override
     public int getId() {
-        return 0x00;
+        return 0x01;
     }
 
     @Override
     public void write(Buffer buffer) {
-
+        buffer.writeLong(this.clientTime);
     }
 
     @Override
     public void read(Buffer buffer) {
-
+        this.clientTime = buffer.readLong();
     }
+
 }
