+++ protocol/src/main/java/org/feathercore/protocol/packet/status/server/S01Pong.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright 2019 Feather Core
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.feathercore.protocol.packet.status.server;
+
+import org.feathercore.protocol.packet.status.client.C01Ping;
+
+/**
+ * Created by k.shandurenko on 09/04/2019
+ */
+public class S01Pong extends C01Ping {
+
+}
